OPENEUROPA-1895: Use interface in proxy ticket middleware.
